feat: User Policy

+++ app/Http/Controllers/Api/V1/UserController.php
@@ -3,6 +3,7 @@
 namespace App\Http\Controllers\Api\V1;
 
 use App\Http\Controllers\Controller;
+use Illuminate\Auth\Access\AuthorizationException;
 use Illuminate\Http\JsonResponse;
 use Illuminate\Support\Facades\Hash;
 use App\Http\Requests\V1\{
@@ -23,9 +24,12 @@ class UserController extends Controller
      * Returns list of all users.
      *
      * @return UserCollection
+     * @throws AuthorizationException
      */
     public function index(): UserCollection
     {
+        $this->authorize('list', auth()->user());
+
         $users = User::all();
 
         return new UserCollection($users);
@@ -35,10 +39,13 @@ class UserController extends Controller
      * Show single User data.
      *
      * @param User $user
-     * @return UserResource
+     * @return JsonResponse|UserResource
+     * @throws AuthorizationException
      */
-    public function show(User $user): UserResource
+    public function show(User $user): JsonResponse|UserResource
     {
+        $this->authorize('isOwner', [$user, 'visualizar esse usuário.']);
+
         return new UserResource($user);
     }
 
@@ -48,9 +55,12 @@ class UserController extends Controller
      * @param UpdateUserRequest $request
      * @param User $user
      * @return JsonResponse
+     * @throws AuthorizationException
      */
     public function update(UpdateUserRequest $request, User $user): JsonResponse
     {
+        $this->authorize('isOwner', [$user, 'atualizar esse usuário.']);
+
         $data = $request->all();
 
         if ($request->password) {
@@ -61,7 +71,7 @@ class UserController extends Controller
             unset($data['password']);
         }
 
-        if(auth()->user()->email !== $request->email) {
+        if (auth()->user()->email !== $request->email) {
             auth()->user()->newEmail($request->email);
             $data['email'] = auth()->user()->email;
         }
@@ -74,12 +84,15 @@ class UserController extends Controller
     /**
      * @param User $user
      * @return JsonResponse
+     * @throws AuthorizationException
      */
     public function destroy(User $user): JsonResponse
     {
+        $this->authorize('isOwner', [$user, 'excluir esse usuário.']);
+
         $user->delete();
 
-        return response()->json(['message' => 'User deleted']);
+        return response()->json(['message' => 'Usuário excluído']);
     }
 
     /**
@@ -87,13 +100,19 @@ class UserController extends Controller
      *
      * @param User $user
      * @return JsonResponse
+     * @throws AuthorizationException
      */
     public function deleteAvatar(User $user): JsonResponse
     {
+        $this->authorize('isOwner', [
+            $user,
+            'excluir a imagem de perfil desse usuário.'
+        ]);
+
         $user->avatar_url = null;
         $user->save();
 
-        return response()->json(['message' => 'Avatar deleted']);
+        return response()->json(['message' => 'Avatar excluído']);
     }
 
     /**
@@ -101,9 +120,18 @@ class UserController extends Controller
      *
      * @param StoreResourceUserRequest $request
      * @return JsonResponse
+     * @throws AuthorizationException
      */
     public function storeResource(StoreResourceUserRequest $request): JsonResponse
     {
+        $this->authorize('isRequestUser',
+            [
+                User::class,
+                $request->userId,
+                'adicionar esse recurso na lista de salvos'
+            ]
+        );
+
         $resourceUser = ResourceUser::create($request->validated());
 
         return response()->json($resourceUser, 201);
@@ -115,9 +143,16 @@ class UserController extends Controller
      * @param User $user
      * @param Resource $resource
      * @return JsonResponse
+     * @throws AuthorizationException
      */
     public function deleteResource(User $user, Resource $resource): JsonResponse
     {
+        $this->authorize('isRequestUser', [
+            User::class,
+            $user->id,
+            'adicionar esse recurso na lista de salvos'
+        ]);
+
         $user->resources()->detach($resource);
 
         return response()->json(['message' => 'Resource deleted']);
